Replace usage of custom describe blocks with test regex patterns
