+++ lib/buttons-and-links/__tests__/KButton.spec.js
@@ -1,9 +1,9 @@
 import { shallowMount } from '@vue/test-utils';
 import KButton from '../KButton.vue';
-import { renderComponent, takeSnapshot } from '../../../jest.conf/testUtils';
+import { renderComponent, takeSnapshot } from '../../../jest.conf/visual.testUtils';
 
 describe('KButton', () => {
-  describeUnit('icon related props', () => {
+  describe('icon related props', () => {
     it('should render an icon before the text with the icon string passed to the `icon` prop', () => {
       const wrapper = shallowMount(KButton, {
         propsData: {
@@ -30,7 +30,7 @@ describe('KButton', () => {
     });
   });
 
-  describeUnit('text prop and slots', () => {
+  describe('text prop and slots', () => {
     it('should render the text prop if nothing is in the default slot', () => {
       const wrapper = shallowMount(KButton, {
         propsData: {
@@ -54,7 +54,7 @@ describe('KButton', () => {
     });
   });
 
-  describeUnit('event handling', () => {
+  describe('event handling', () => {
     it('should emit a click event when clicked', () => {
       const wrapper = shallowMount(KButton, {
         propsData: {
@@ -66,10 +66,10 @@ describe('KButton', () => {
     });
   });
 
-  describeVisual('KButton Visual Tests', () => {
+  describe('[Visual] KButton Visual Tests', () => {
     it('renders correctly with default props', async () => {
-      await renderComponent(page, 'KButton', { text: 'Test Button' });
-      await takeSnapshot(page, 'KButton - Default');
+      await renderComponent('KButton', { text: 'Test Button' });
+      await takeSnapshot('KButton - Default');
     });
   });
 });
